- Implemented CancellationToken in services and repositories - Implemented UnitOfWork pattern for transactions and saving data - Made ShopSaleService.cs and IShopSaleService.cs - Made refactoring - Made fixes and improvements

+++ src/BShop/BShop.Infrastructure/Repositories/ProductRepository.cs
@@ -6,33 +6,38 @@ namespace BShop.Infrastructure.Repositories;
 
 public class ProductRepository(ShopDbContext context) : IProductRepository
 {
-    public async Task<Product?> GetProductById(Guid id)
+    public async Task<Product?> GetProductById(Guid id, CancellationToken cancellationToken)
     {
-        return await context.Products.FindAsync(id);
+        return await context.Products.FirstOrDefaultAsync(x => x.Id == id, cancellationToken);
     }
 
-    public async Task<Product> CreateProduct(Product product)
+    public async Task CreateProduct(Product product, CancellationToken cancellationToken)
     {
         context.Products.Add(product);
-        await context.SaveChangesAsync();
-        return product;
+        await context.SaveChangesAsync(cancellationToken);
     }
 
-    public async Task UpdateProduct(Product product)
+    public async Task UpdateProduct(Product product, CancellationToken cancellationToken)
     {
         context.Products.Update(product);
-        await context.SaveChangesAsync();
+        await context.SaveChangesAsync(cancellationToken);
     }
 
-    public async Task DeleteProduct(Guid id)
+    public async Task DeleteProduct(Guid id, CancellationToken cancellationToken)
     {
-        var product = await GetProductById(id);
-        context.Products.Remove(product);
-        await context.SaveChangesAsync();
+        var product = await GetProductById(id, cancellationToken);
+
+        if (product != null)
+        {
+            context.Products.Remove(product);
+            await context.SaveChangesAsync(cancellationToken);
+        }
     }
 
-    public IQueryable<Product> GetAllProducts()
+    public IQueryable<Product> GetAllProducts(CancellationToken cancellationToken)
     {
-        return context.Products;
+        return context.Products
+            .AsNoTracking()
+            .Where(x => !x.IsDeleted);
     }
 }
